requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
flask
langchain_experimental
GitPython
chromadb
tiktoken